refactor(plugins): :truck: rename shiki transformer files
