+++ src/plugins/shiki-transformers.ts
@@ -1,7 +1,10 @@
 import { h } from 'hastscript'
 import type { ShikiTransformer } from 'shiki'
 
-export { transformerNotationDiff, transformerNotationHighlight } from './shikiOfficialTransformers'
+export {
+  transformerNotationDiff,
+  transformerNotationHighlight
+} from './shiki-official-transformers'
 
 function parseMetaString(str = '') {
   return Object.fromEntries(
